+++ TemplateComponent/TemplateComponent.test.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { fireEvent, screen, render } from "../../test-setup";
+import { TemplateComponent } from "./TemplateComponent";
+
+describe("Tests for the TemplateComponent component", () => {
+  it("should render", () => {
+    render(<TemplateComponent />);
+  });
+});